+++ packages/terra-overlay/CHANGELOG.md
@@ -7,6 +7,7 @@ Unreleased
 * Added `data-overlay-count` attribute to root document when overlay is opened
 * Added check for `data-modal-count` attribute on root document to track opening and closing of overlays and modals,
   to prevent the removal of the `inert` attribute from the root element if another overlay or abstract modal exists.
+* Added Node.contains polyfill to better support IE 10 / IE 11
 
 3.17.0 - (June 28, 2019)
 ------------------